# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maintenance Management API",
        "description": "API for managing maintenance products and records with different user roles (admin, customer, technician)",
        "contact": {
            "name": "API Support",
            "email": "support@maintenance-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/technician/assigned_products": {
            "get": {
                "tags": [
                    "Technician"
                ],
                "summary": "Get technician's assigned products",
                "description": "Retrieve products assigned to the authenticated technician",
                "operationId": "380cef8bc67ebce0561471cb5a32041c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/technician/maintenance_records": {
            "get": {
                "tags": [
                    "Technician"
                ],
                "summary": "Get technician's maintenance records",
                "description": "Retrieve maintenance records assigned to the authenticated technician",
                "operationId": "01d83b71c237fbad09f896018b7efb3b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MaintenanceRecord"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/technician/maintenance_records/{id}": {
            "put": {
                "tags": [
                    "Technician"
                ],
                "summary": "Update maintenance record",
                "description": "Update a maintenance record assigned to the authenticated technician",
                "operationId": "f62437c865c61f5101853d9858797431",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of maintenance record to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaintenanceRecord"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Maintenance record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaintenanceRecord"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Maintenance record not assigned to technician"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/technician/products/{id}": {
            "get": {
                "tags": [
                    "Technician"
                ],
                "summary": "Get assigned product details",
                "description": "Retrieve details of a product assigned to the authenticated technician",
                "operationId": "e457b9885b37f957152af2bbef499e56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/technician/products/{id}/maintenance-history": {
            "get": {
                "tags": [
                    "Technician"
                ],
                "summary": "Get product maintenance history",
                "description": "Retrieve maintenance history for a product assigned to the authenticated technician",
                "operationId": "4dd56da1ffde5cddc9ad5f0b73fd684b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MaintenanceRecord"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "enum": [
                            "Fire Extinguisher",
                            "Smoke Detector",
                            "Fire Alarm"
                        ]
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "water",
                            "foam",
                            "CO2",
                            "DCP"
                        ]
                    },
                    "type_capacity": {
                        "type": "string"
                    },
                    "serial_number": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Active",
                            "Expired",
                            "Needs Maintenance"
                        ]
                    },
                    "location": {
                        "type": "string"
                    },
                    "assigned_to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MaintenanceRecord": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "technician_id": {
                        "type": "integer"
                    },
                    "maintenance_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "completed",
                            "pending",
                            "overdue"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "oauth2",
                "description": "OAuth2 Bearer Token for authentication",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "scopes": {
                            "read": "Read access",
                            "write": "Write access"
                        }
                    }
                }
            }
        }
    },
...
}